fix: 함수 분리 및 파일 명 변경
